refactor(Gradle): Move top-level functions to utility objects
This is a prerequisite for leveraging Gradle's configuration cache which
cannot handle top-level functions [1].

[1]: https://docs.gradle.org/8.2/userguide/configuration_cache.html#config_cache:not_yet_implemented:accessing_top_level_at_execution
