+++ build.gradle.kts
@@ -89,83 +89,9 @@ tasks.register("allDependencies") {
     }
 }
 
-val copyrightExcludedPaths = listOf(
-    "LICENSE",
-    "NOTICE",
-    "batect",
-    "gradlew",
-    "gradle/",
-    "examples/",
-    "integrations/completions/",
-    "plugins/reporters/asciidoc/src/main/resources/pdf-theme/pdf-theme.yml",
-    "plugins/reporters/asciidoc/src/main/resources/templates/freemarker_implicit.ftl",
-    "plugins/reporters/fossid/src/main/resources/templates/freemarker_implicit.ftl",
-    "plugins/reporters/freemarker/src/main/resources/templates/freemarker_implicit.ftl",
-    "plugins/reporters/static-html/src/main/resources/prismjs/",
-    "plugins/reporters/web-app-template/yarn.lock",
-    "resources/META-INF/",
-    "resources/exceptions/",
-    "resources/licenses/",
-    "resources/licenserefs/",
-    "test/assets/",
-    "funTest/assets/"
-)
-
-val copyrightExcludedExtensions = listOf(
-    "css",
-    "graphql",
-    "json",
-    "md",
-    "png",
-    "svg",
-    "ttf"
-)
-
-fun getCopyrightableFiles(rootDir: File): List<File> {
-    val gitFilesProvider = providers.of(GitFilesValueSource::class) { parameters { workingDir = rootDir } }
-
-    return gitFilesProvider.get().filter { file ->
-        val isHidden = file.toPath().any { it.toString().startsWith(".") }
-
-        !isHidden
-                && copyrightExcludedPaths.none { it in file.invariantSeparatorsPath }
-                && file.extension !in copyrightExcludedExtensions
-    }
-}
-
-val maxCopyrightLines = 50
-
-fun extractCopyrights(file: File): List<String> {
-    val copyrights = mutableListOf<String>()
-
-    var lineCounter = 0
-
-    file.useLines { lines ->
-        lines.forEach { line ->
-            if (++lineCounter > maxCopyrightLines) return@forEach
-            val copyright = line.replaceBefore(" Copyright ", "", "").trim()
-            if (copyright.isNotEmpty() && !copyright.endsWith("\"")) copyrights += copyright
-        }
-    }
-
-    return copyrights
-}
-
-val copyrightPrefixRegex = Regex("Copyright .*\\d{2,}(-\\d{2,})? ", RegexOption.IGNORE_CASE)
-
-fun extractCopyrightHolders(statements: Collection<String>): List<String> {
-    val holders = mutableListOf<String>()
-
-    statements.mapNotNullTo(holders) { statement ->
-        val holder = statement.replace(copyrightPrefixRegex, "")
-        holder.takeUnless { it == statement }?.trim()
-    }
-
-    return holders
-}
-
 val checkCopyrightsInNoticeFile by tasks.registering {
-    val files = getCopyrightableFiles(rootDir)
+    val gitFilesProvider = providers.of(GitFilesValueSource::class) { parameters { workingDir = rootDir } }
+    val files = CopyrightableFiles.filter(gitFilesProvider)
     val noticeFile = rootDir.resolve("NOTICE")
     val genericHolderPrefix = "The ORT Project Authors"
 
@@ -176,7 +102,7 @@ val checkCopyrightsInNoticeFile by tasks.registering {
         var hasViolations = false
 
         files.forEach { file ->
-            val copyrights = extractCopyrights(file)
+            val copyrights = CopyrightUtils.extract(file)
             if (copyrights.isNotEmpty()) {
                 allCopyrights += copyrights
             } else {
@@ -186,7 +112,7 @@ val checkCopyrightsInNoticeFile by tasks.registering {
         }
 
         val notices = noticeFile.readLines()
-        extractCopyrightHolders(allCopyrights).forEach { holder ->
+        CopyrightUtils.extractHolders(allCopyrights).forEach { holder ->
             if (!holder.startsWith(genericHolderPrefix) && notices.none { holder in it }) {
                 hasViolations = true
                 logger.error("The '$holder' Copyright holder is not captured in '$noticeFile'.")
@@ -197,51 +123,9 @@ val checkCopyrightsInNoticeFile by tasks.registering {
     }
 }
 
-val lastLicenseHeaderLine = "License-Filename: LICENSE"
-
-val expectedCopyrightHolder =
-    "The ORT Project Authors (see <https://github.com/oss-review-toolkit/ort/blob/main/NOTICE>)"
-
-// The header without `lastLicenseHeaderLine` as that line is used as a marker.
-val expectedLicenseHeader = """
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        https://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
-    SPDX-License-Identifier: Apache-2.0
-""".trimIndent()
-
-fun extractLicenseHeader(file: File): List<String> {
-    var headerLines = file.useLines { lines ->
-        lines.takeWhile { !it.endsWith(lastLicenseHeaderLine) }.toList()
-    }
-
-    while (true) {
-        val uniqueColumnChars = headerLines.mapNotNullTo(mutableSetOf()) { it.firstOrNull() }
-
-        // If there are very few different chars in a column, assume that column to consist of comment chars /
-        // indentation only.
-        if (uniqueColumnChars.size < 3) {
-            val trimmedHeaderLines = headerLines.mapTo(mutableListOf()) { it.drop(1) }
-            headerLines = trimmedHeaderLines
-        } else {
-            break
-        }
-    }
-
-    return headerLines
-}
-
 val checkLicenseHeaders by tasks.registering {
-    val files = getCopyrightableFiles(rootDir)
+    val gitFilesProvider = providers.of(GitFilesValueSource::class) { parameters { workingDir = rootDir } }
+    val files = CopyrightableFiles.filter(gitFilesProvider)
 
     inputs.files(files)
 
@@ -252,15 +136,15 @@ val checkLicenseHeaders by tasks.registering {
         var hasErrors = false
 
         files.forEach { file ->
-            val headerLines = extractLicenseHeader(file)
+            val headerLines = LicenseUtils.extractHeader(file)
 
-            val holders = extractCopyrightHolders(headerLines)
-            if (holders.singleOrNull() != expectedCopyrightHolder) {
+            val holders = CopyrightUtils.extractHolders(headerLines)
+            if (holders.singleOrNull() != CopyrightUtils.expectedHolder) {
                 hasErrors = true
                 logger.error("Unexpected copyright holder(s) in file '$file': $holders")
             }
 
-            if (!headerLines.joinToString("\n").endsWith(expectedLicenseHeader)) {
+            if (!headerLines.joinToString("\n").endsWith(LicenseUtils.expectedHeader)) {
                 hasErrors = true
                 logger.error("Unexpected license header in file '$file'.")
             }
